+++ src/ast/visitors/print_visitor.h
@@ -166,10 +166,10 @@ class PrintVisitor : public AST::Visitor {
         [this](const Lit& lit) {
           std::visit(
             overloaded {
-                [&](const int i) { print("Lit Int: " + std::to_string(i), lit.id); },
+                [&](const int64_t i) { print("Lit Int: " + std::to_string(i), lit.id); },
                 [&](const std::string& s) { print("Lit String: " + s, lit.id); },
                 [&](const bool& b) { print("Lit Bool: " + std::to_string(b), lit.id); },
-                [&](const float& b) { print("Lit Float: " + std::to_string(b), lit.id); }
+                [&](const double& b) { print("Lit Float: " + std::to_string(b), lit.id); }
             }, lit.kind);
         },
         [this](const P<Block>& block) {
